+++ src/CADL.Extension/Emitter.Csharp/src/lib/ServiceAuthentication.ts
@@ -1,0 +1,44 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+import { ServiceAuthentication } from "@cadl-lang/rest/http";
+import { InputApiKeyAuth } from "../type/InputApiKeyAuth.js";
+import { InputAuth } from "../type/InputAuth.js";
+import { InputOAuth2Auth } from "../type/InputOAuth2Auth.js";
+
+export function processServiceAuthentication(
+    authentication: ServiceAuthentication
+): InputAuth {
+    const auth = {} as InputAuth;
+    let scopes: Set<string> | undefined;
+
+    for (const option of authentication.options) {
+        for (const schema of option.schemes) {
+            switch (schema.type) {
+                case "apiKey":
+                    auth.ApiKey = { Name: schema.name } as InputApiKeyAuth;
+                    break;
+                case "oauth2":
+                    for (const flow of schema.flows) {
+                        if (flow.scopes) {
+                            scopes ??= new Set<string>();
+                            for (const scope of flow.scopes) {
+                                scopes.add(scope.value);
+                            }
+                        }
+                    }
+                    break;
+                default:
+                    throw new Error("Not supported authentication.");
+            }
+        }
+    }
+
+    if (scopes) {
+        auth.OAuth2 = {
+            Scopes: Array.from(scopes.values())
+        } as InputOAuth2Auth;
+    }
+
+    return auth;
+}
